tok: fix overflow for "mute music only"
Also introducing a new attribute "translation_long" for a translation
that is more precise but exceeds the limit. Verified that the game
ignores it.

If a conversion tool is made to convert these translations to sitelen
pona, it can prefer the `translation_long` fields when present, since
they'll fit. (Unless they're *very* long)